refactor imports; fix flake8 issues
