+++ pyfastapi/controllers/__init__.py
@@ -1,0 +1,5 @@
+from .continent import router as continent_router
+from .country import router as country_router
+from .person import router as person_router
+
+__all__ = ["continent_router", "country_router", "person_router"]
