+++ pyfastapi/libs/__init__.py
@@ -1,0 +1,3 @@
+from .db import get_db, Base
+
+__all__ = ["get_db", "Base"]
