+++ pyfastapi/main.py
@@ -1,9 +1,8 @@
 from fastapi import FastAPI
 from fastapi_pagination import add_pagination
 
-from pyfastapi.controllers.person import router as person_router
-from pyfastapi.controllers.country import router as country_router
-from pyfastapi.controllers.continent import router as continent_router
+from pyfastapi.controllers import person_router, country_router, continent_router
+
 
 app = FastAPI()
 
@@ -13,10 +12,9 @@ app.include_router(country_router, prefix="/countries")
 app.include_router(continent_router, prefix="/continents")
 app.add_api_route(
     "/health",
-    endpoint=lambda : {"status": "ok"},
+    endpoint=lambda: {"status": "ok"},
     methods=["GET"]
 )
 
 
 app = add_pagination(app)
-
