+++ pyfastapi/models/__init__.py
@@ -1,0 +1,5 @@
+from .continent import Continent
+from .country import Country
+from .person import Person
+
+__all__ = ["Continent", "Country", "Person"]
