+++ pyfastapi/repositories/__init__.py
@@ -1,0 +1,6 @@
+from .continent import ContinentRepository
+from .country import CountryRepository
+from .person import PersonRepository
+
+
+__all__ = ["ContinentRepository", "CountryRepository", "PersonRepository"]
