+++ pyfastapi/repositories/base.py
@@ -2,7 +2,7 @@ from abc import ABC
 from fastapi import Depends
 from sqlalchemy.orm import Session
 
-from pyfastapi.libs.db import get_db
+from pyfastapi.libs import get_db
 
 
 class BaseRepository(ABC):
