+++ pyfastapi/schemas/__init__.py
@@ -1,0 +1,5 @@
+from .continent import ContinentSchema
+from .country import CountrySchema, CountryListSchema
+from .person import PersonListSchema, PersonCreateSchema
+
+__all__ = ["ContinentSchema", "CountrySchema", "CountryListSchema", "PersonListSchema", "PersonCreateSchema"]
